+++ pysrc/whenever/_core.py
@@ -48,6 +48,7 @@ except ModuleNotFoundError as e:
         _unpkl_tdelta,
         _unpkl_time,
         _unpkl_utc,
+        _unpkl_vdelta,
         _unpkl_ym,
         _unpkl_zoned,
     )